chore: Removed unused resources
