--- src/main/resources/GUI/CSS/Tabs.css
@@ -1,120 +1,0 @@
-
-
-* {
-    -fx-primary-color: rgb(102,106,106);
-    -fx-secondary-color: rgb(102,106,106);
-    -fx-focus-color: -fx-secondary-color;
-}
-
-.tab-header-area .tab-header-background {
-    -fx-background-color: -fx-primary-color;
-}
-.tab-pane .tab:selected {
-    -fx-background-color: -fx-secondary-color;
-    -fx-background-insets: 0 0 0 0;
-    -fx-background-radius: 0 0 0 0;
-}
-.standard-tab .tab-label {
-    -fx-alignment: CENTER;
-    -fx-background-color: transparent;
-    -fx-text-fill: darkblue;
-    -fx-font-weight: bold;
-    -fx-font-size: 19;
-    -fx-padding: 0 10 0 .1;
-}
-
-
-.error-tab .tab-label {
-    -fx-alignment: CENTER;
-    -fx-background-color: transparent;
-    -fx-text-fill: rgb(100,0,0);
-    -fx-font-weight: bold;
-    -fx-font-size: 19;
-    -fx-padding: 0 .1 0 10;
-}
-
-
-
-
-.tab-pane:top *.tab-header-area {
-    -fx-padding: 0 0 0 10;
-
-    /* NO SPACE LEFT OF FIRST TAB */
-    /* -fx-padding: 0 0 0 30; */
-}
-.tab-pane {
-    -fx-background-color: transparent;
-    -fx-tab-min-width: 50px;
-    -fx-tab-min-height: 55px;
-    -fx-tab-max-height: 55px;
-}
-.tab-pane .tab {
-    -fx-background-color: transparent;
-}
-.tab-pane:focused .tab-header-area .headers-region .tab:selected .focus-indicator {
-    -fx-border-width: 5;
-    -fx-border-insets: 0 0 0 0;
-    -fx-border-radius: 0;
-}
-
-/*
-.tab-pane:top .tab-header-area .tab-header-background {
-    -fx-background-color: rgb(102,105,106);
-}
-
-.tab:selected:top .tab-container .tab-label {
-    -fx-text-fill: lightblue;
-}
-*/
-
-/*
-.tab-standard .tab:selected:top .tab-container .tab-label {
-    -fx-background-color: transparent;
-    -fx-text-fill: red;
-}
-
-.tab-standard .tab-pane:top .tab-header-area .headers-region .tab:top .tab-container {
-    -fx-background-color: red;
-}
-
-.tab-error .tab:selected:top .tab-container .tab-label {
-    -fx-background-color: transparent;
-    -fx-text-fill: red;
-}
-*/
-
-/*
-#tab-standard .tab-pane:top .tab-header-area,
-#tab-standard .tab-pane:top .tab-header-area .headers-region,
-#tab-standard .tab-pane:top .tab-header-area .headers-region .tab:selected:top,
-#tab-standard .tab-pane:top .tab-header-area .headers-region .tab:selected:top .tab-container,
-#tab-standard .tab-pane:top .tab-header-area .headers-region .tab:selected:top .tab-container .focus-indicator {
-    -fx-background-color: transparent;
-    -fx-text-fill: darkblue;
-}
-
-#tab-standard .tab:selected:top .tab-container .tab-label {
-    -fx-background-color: transparent;
-    -fx-text-fill: darkblue;
-}
-#tab-standard .tab-pane:top .tab-header-area .tab-header-background {
-    -fx-background-color: transparent;
-    -fx-text-fill: darkblue;
-}
-
-#tab-standard .tab:selected:top .tab-container .tab-label {
-    -fx-text-fill: lightblue;
-}
-
-
-#tab-error .tab-pane:top .tab-header-area,
-#tab-error .tab-pane:top .tab-header-area .headers-region,
-#tab-error .tab-pane:top .tab-header-area .headers-region .tab:selected:top,
-#tab-error .tab-pane:top .tab-header-area .headers-region .tab:selected:top .tab-container,
-#tab-error .tab-pane:top .tab-header-area .headers-region .tab:selected:top .tab-container .focus-indicator,
-#tab-error .tab:selected:top .tab-container .tab-label,
-#tab-error .tab-pane:top .tab-header-area .tab-header-background {
-    -fx-background-color: transparent;
-    -fx-text-fill: darkred;
-}
-*/
